+++ rust/operator-binary/src/controller/validate.rs
@@ -26,14 +26,13 @@ use stackable_operator::{
 };
 use strum::{EnumDiscriminants, IntoEnumIterator, IntoStaticStr};
 
-use super::{
-    ValidatedCluster, ValidatedClusterConfig, ValidatedRoleConfig, ValidatedTls,
-    ValidatedTrinoConfig,
-};
 use crate::{
     authentication::{self, TrinoAuthenticationConfig, TrinoAuthenticationTypes},
-    controller::dereference::DereferencedObjects,
-    crd::{Container, TrinoRole, v1alpha1},
+    controller::{
+        ValidatedCluster, ValidatedClusterConfig, ValidatedRoleConfig, ValidatedTls,
+        ValidatedTrinoConfig, dereference::DereferencedObjects,
+    },
+    crd::{Container, TrinoRole, catalog::TrinoCatalogName, v1alpha1},
 };
 
 #[derive(Snafu, Debug, EnumDiscriminants)]
@@ -109,6 +108,11 @@ pub enum Error {
         "the Vector aggregator discovery ConfigMap name is required when the Vector agent is enabled"
     ))]
     MissingVectorAggregatorConfigMapName,
+
+    #[snafu(display(
+        "The catalog name {catalog_name:?} clashes (there are multiple catalogs with this name). Please make sure there is exactly one catalog for any given name"
+    ))]
+    ClashingCatalogName { catalog_name: TrinoCatalogName },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
@@ -271,6 +275,19 @@ pub fn validate(
         role_group_configs.insert(trino_role, groups);
     }
 
+    let mut catalogs = BTreeMap::new();
+    for catalog in &dereferenced_objects.catalogs {
+        if catalogs
+            .insert(catalog.name.clone(), catalog.clone())
+            .is_some()
+        {
+            return ClashingCatalogNameSnafu {
+                catalog_name: catalog.name.clone(),
+            }
+            .fail();
+        }
+    }
+
     let tls = &trino.spec.cluster_config.tls;
     let cluster_config = ValidatedClusterConfig {
         tls: ValidatedTls {
@@ -282,7 +299,7 @@ pub fn validate(
         fault_tolerant_execution: dereferenced_objects.resolved_fte_config.clone(),
         client_protocol: dereferenced_objects.resolved_client_protocol_config.clone(),
         coordinator_pod_refs: trino.coordinator_pods(&namespace).collect(),
-        catalogs: dereferenced_objects.catalogs.clone(),
+        catalogs,
     };
 
     let name = get_cluster_name(trino).context(GetClusterNameSnafu)?;